Delete Superfluid CL migration code
